django-fastapi/micro-service/pyproject.toml
[project]
name = "micro-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=44.0.1",
    "fastapi>=0.115.8",
    "pyjwt>=2.10.1",
    "uvicorn>=0.34.0",
    "websockets>=14.2",
]
